requirements.txt
openai>=1.0.0
paramiko>=3.0.0
pyyaml>=6.0
pydantic>=2.0.0
